refactor for easier maintainance (#160)
